blaster/requirements.txt
beautifulsoup4==4.3.2
nltk==3.0.2
requests=2.7.0
newspaper=0.1.5
enchant
spacy=0.8
textblob=0.9.0
redis=2.10.3
elasticsearch>=1.0.0,<2.0.0